Drop containers package, move to tool_util
