+++ packages/config/requirements.txt
@@ -1,4 +1,3 @@
-galaxy-containers
 boltons
 pykwalify
 PyYAML
