--- packages/containers/HISTORY.rst
@@ -1,30 +1,0 @@
-.. :changelog:
-
-History
--------
-
-.. to_doc
-
----------------------
-21.1.1.dev0
----------------------
-
-
-
----------------------
-21.1.0 (2021-03-19)
----------------------
-
-* First release from the 21.01 branch of Galaxy.
-
----------------------
-20.5.0 (2020-07-04)
----------------------
-
-* First release from the 20.05 branch of Galaxy.
-
----------------------
-19.9.0 (2019-12-16)
----------------------
-
-* Initial import from dev branch of Galaxy during 19.09 development cycle.
