--- packages/containers/Makefile
@@ -1,1 +1,0 @@
-../package.Makefile
