--- packages/containers/README.rst
@@ -1,14 +1,0 @@
-
-.. image:: https://badge.fury.io/py/galaxy-containers.svg
-   :target: https://pypi.org/project/galaxy-containers/
-
-
-Overview
---------
-
-The Galaxy_ containers module - interfaces for modeling and interacting with container backends (docker and docker swarm currently).
-
-* Free software: Academic Free License version 3.0
-* Code: https://github.com/galaxyproject/galaxy
-
-.. _Galaxy: http://galaxyproject.org/
