--- packages/containers/dev-requirements.txt
@@ -1,1 +1,0 @@
-../package-dev-requirements.txt
