--- packages/containers/galaxy/containers
@@ -1,1 +1,0 @@
-../../../lib/galaxy/containers
