--- packages/containers/galaxy/project_galaxy_containers.py
@@ -1,9 +1,0 @@
-__version__ = "22.5.0.dev0"
-
-PROJECT_NAME = "galaxy-containers"
-PROJECT_OWNER = PROJECT_USERAME = "galaxyproject"
-PROJECT_URL = "https://github.com/galaxyproject/galaxy"
-PROJECT_AUTHOR = "Galaxy Project and Community"
-PROJECT_DESCRIPTION = "Galaxy Container Modeling and Interaction Abstractions"
-PROJECT_EMAIL = "galaxy-committers@lists.galaxyproject.org"
-RAW_CONTENT_URL = f"https://raw.github.com/{PROJECT_USERAME}/{PROJECT_NAME}/master/"
