--- packages/containers/requirements.txt
@@ -1,3 +1,0 @@
-galaxy-util
-requests
-
