--- packages/containers/scripts
@@ -1,1 +1,0 @@
-../build_scripts
