--- packages/containers/setup.cfg
@@ -1,1 +1,0 @@
-../../setup.cfg
