--- packages/containers/tests
@@ -1,1 +1,0 @@
-../../test/unit/containers
