+++ packages/tool_util/requirements.txt
@@ -1,5 +1,4 @@
 galaxy-util>=20.1.0.dev0
-galaxy-containers
 lxml
 pydantic
 pyyaml
